REF: move everything back to Series/DataFrame classes
