+++ pandas/core/ops/__init__.py
@@ -9,6 +9,7 @@ from pandas.core.ops.array_ops import (
     arithmetic_op,
     comp_method_OBJECT_ARRAY,
     comparison_op,
+    fill_binop,
     get_array_op,
     logical_op,
     maybe_prepare_scalar_for_op,
@@ -17,6 +18,7 @@ from pandas.core.ops.common import (
     get_op_result_name,
     unpack_zerodim_and_defer,
 )
+from pandas.core.ops.docstrings import make_flex_doc
 from pandas.core.ops.invalid import invalid_comparison
 from pandas.core.ops.mask_ops import (
     kleene_and,
@@ -70,10 +72,12 @@ __all__ = [
     "comparison_op",
     "comp_method_OBJECT_ARRAY",
     "invalid_comparison",
+    "fill_binop",
     "kleene_and",
     "kleene_or",
     "kleene_xor",
     "logical_op",
+    "make_flex_doc",
     "radd",
     "rand_",
     "rdiv",
